LayoutTests/platform/wpe/imported/w3c/web-platform-tests/css/css-ui/user-select-001-expected.txt

Let's select this word

FAIL Select the text 'word' assert_equals: The text 'word' should be selectable. expected "word" but got ""
